Add footnotes for mobile view
It turns out it's hard to find out that you can click on the hidden side
notes when on mobile.  As a fallback I am generating the footnotes in
the org export in addition to the sidenotes.  This way if a user scrolls
down to find the footnotes they are there.  On big monitors they aren't
needed however since the sidenotes are shown by default.