start at current file with grep-find
Instead of always going to the earliest or latest result, this makes
sure to move the cursor to the current match.  This is intended to be
used with embark, so that it searches for a symbol that exists in a
file.  The effect then is that it does a grep search which can be
traversed with `M-x next-error`, and the search starts from the point it
was called.  Then the user can go back and forth relative to where they
started.